+++ src/components/tabs/tab/Tab.svelte
@@ -6,6 +6,7 @@
   import { type StringBoolean } from "../../../common/types/StringBoolean";
   import { type TabSize } from "../types/TabSize";
 
+  export let label: string = "";
   export let href: string = "";
   export let title: string = "";
   export let active: StringBoolean = "false";
@@ -20,7 +21,7 @@
   aria-selected={active === "true" ? "true" : "false"}
 >
   <a {href} {title} class="link {active === 'true' ? 'link-active' : ''}">
-    <slot>Tab</slot>
+    {label}
   </a>
 </li>
 
@@ -56,6 +57,7 @@
   }
 
   .tab {
+    background-color: var(--khao-sys-color-surface-variantd10);
     list-style-type: none;
     display: inline-flex;
     flex-direction: row;
@@ -76,17 +78,19 @@
 
     &:hover {
       text-decoration: none !important;
-      color: color-mix(
-        in srgb,
-        var(--khao-tab-text-color),
-        var(--khao-tab-layer-color) var(--khao-tab-hover-state-layer-opacity)
-      );
+      background-color: var(--khao-sys-color-outline);
+    }
+
+    &:focus-within {
+      text-decoration: none !important;
+      background-color: var(--khao-sys-color-outline);
     }
   }
 
   .tab-active {
     border-bottom: none;
     height: var(--khao-tab-active-height);
+    background-color: transparent;
   }
 
   .tab-size-flex {
@@ -129,18 +133,8 @@
       );
     }
 
-    &:focus {
-      color: color-mix(
-        in srgb,
-        var(--khao-tab-text-color),
-        var(--khao-tab-layer-color) var(--khao-tab-focus-state-layer-opacity)
-      );
-      background-color: color-mix(
-        in srgb,
-        var(--khao-tab-text-color),
-        var(--khao-tab-layer-color) var(--khao-tab-focus-state-layer-opacity)
-      );
-      font-weight: 800;
+    &:focus-visible {
+      outline: none;
     }
   }