+++ CHANGELOG.md
@@ -5,9 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.3]
+
+- added focus state for `khao-video`
+- removed underline for focus state of secondary Link in `khao-link`
+
 ## [1.9.2]
 
 - upgraded to `khao-malet@0.0.19`
+- added `label` prop to `khao-link`, slot is deprecated now!
+- improved focus state of `khao-link`
+- added bg-color hover/focus and active state to `khao-tab`
+- replace slot in `khao-tab` by `label`
 
 ## [1.9.1]